refactor(api): split devices.go (1271 → 351 lines) by concern
api/devices.go mixed handlers with all the supporting types, validators,
response builders, and mutation helpers in one file. Same-package split,
no behaviour or signature changes.

  devices.go (slim)        HTTP handlers: List / Get / Create / Update / Delete / Clone
  devices_types.go         all request + response struct types
  devices_validation.go    hostname + YAML input validation + sentinel errors
  devices_responses.go     deviceToResponse + per-protocol response builders
  devices_helpers.go       device creation, mutation, deep-copy, MAC/IP parsing,
                           YAML serialization, saveConfig

go test ./internal/api/... passes; golangci-lint clean.
